JavaScript refactoring: page quality buttons.
I have replaced two switches with one simple loop. The change
reduces the amount of the code improving readability.

There is one new feature here, the checkboxes have now title
attribute, so the user does not have to click on the checkbox
to find out what it means.

Tpt: rebase of this commit + appling Jdlrobson comments.

Change-Id: I76d8f95095f6ac105d5dfa99c7dd472378b93e30