(bug 36070) Add of support of non-arabic page number in page namespace.
Pages of multi-pages books can now have a name with number in their own language.

The pagelist tag links to pages with a localized number but, if the localized page
doesn't exist and a page with an arabic number exist, pagelist will link to this page.
The pages tag and the navigation inside of the page namespace use the same system.

With that system existing wikis won't be broken and they will have time to migrate.

Change-Id: I72e5f07977dcf991124252754f92c71fafe1e151