(bug 37419) Rewriting of Index pages edit form in PHP
Rewriting of the index form in PHP in order to improve performance (the edit toolbars were loaded and showed and then Prp created the form) and allow easy editing of index pages
Creation of json configuration message proofreadpage_index_data_config that replace proofreadpage_index_attributes and proofreadpage_js_attributes.
Wiki can keep using the two old messages. The description of the syntax is done in ProofreadIndexPage::getDataConfig()
This creation allow to add new features in the form:
- list of possible values for a field (show as a select html element)
- help messages
- type for each data
- default values
There is no validation in this form in order to support existant data.
The names and id of field has change: a prefix "wpprpindex-" is added in order to avoid any id conflict

Some generic classes where added:
- ProofreadIndexPage that offer methods in order to get index page content.
- ProofreadIndexEntry that manage an entry in the index page and is configuration.
Use of these two classes inside of existing parts of the ProofreeadPage extension will be done in an other commit.

This change is the same as https://gerrit.wikimedia.org/r/#/c/10741/ but have been duplicate because of a bad manipulation.

Patch set 3: Fix a bug on page creation.
Patch set 4: Rebase + replace of wfMsg*

Change-Id: Idc00bddac8abc58e0b6414bdd9dc3f32cab3337a