(bug 37483) Create index and page namespaces automatically if not set
This is retrocompatible with old installations if namespace name are
ones that are stored in proofreadPageNamespaceNames file.

The namespace names used by the Wikisources are all in this file so I
hope their won't be any bugs for Wikisourc For more security we have
to wait that https://gerrit.wikimedia.org/r/#/c/20876/ is merge in
Wikimedia config before adding this patch into core.

If an other namespace name have been used an error message is display
asking to update the configuration.

Namespaces can be customized with:

$wgProofreadPageNamespaceIds = array(
	'index' => /*INDEX NAMESPACE ID */,
	'page' => /* PAGE NAMESPACE ID */
);

Default ids are 250 for page and 252 for index.

Patch set 2: add of some languages.

Patch set 3: rebase + cleaning of the code use of hook, ability to set
             $pfpgNamespaceIds after inclusion of prp, add
			 of more comments...)

Patch set 4: rename of the global var in order to use the same system
             as the other extensions.
Patch set 5: Minor changes + rebase.

Patch set 6: Hope that fix a register_globals set vulnerability.

Patch set 7: Rebase and merge conflict resolved.

Patch set 8: Tweak line 1 of commit message to be brief and concise.

Patch set 10: Fix Polish translation of the namespace name.

Change-Id: Iada3840561a68113985d778db731a3ef8af523e2