Make Special:IndexPages not show invalid page ids
Special:IndexPages outputs a blank line for anything that's not
a valid title. The query figures out page name and namespace by
outer joining on page. Given that all entries missing page_title
are going to be a blank line anyways, use an inner join instead
of an outer join.

Bug: 32590
Change-Id: I778ecaad2b062795f10506d506c26e492df90079