Remove PageTarget.init, instead load via VisualEditorPluginModules
This avoids the instantiation of the desktop target ahead of the code
for ProofreadPage being loaded, and so a race condition on whether an
editor would load at all (or if the user would just get an error).

Unfortunately, this means that DesktopArticleTarget now always loads
the ProofreadPage integration on every load on desktop (but not yet
on mobile as ext.proofreadpage.ve.pageTarget is desktop only for now).

Bug: T138644
Change-Id: Ic4cbb756033523a537d9e6e2bb4b84cf52ce4369