Fix the ProofreadPage’s VisualEditor
The ProofreadPage’s VisualEditor was broken because of the commit
Ic8ea89a42ad6d873 (cb081bc3c8c3115).

Thanks to Tim’s help. In some cases, like during the API call for
the VisualEditor, the 'parser clone' clones a StubObject because
the real Parser is not already initialised; with unstub() it
becomes initialised.

Change-Id: Iac9f14815dab092bb06e5743a793abfd4aa2b0f5