Uses the Index: page based on file name when there are multiple possible Index: pages
When more than one Index: page links to a Page: page we should use as Index: for this page the Index: pages that shares the same file name, if it exists.

E.g. if Index:Foo and Index:Bar.djvu both link to Page:Bar.djvu/3, we should use as index Index:Bar.djvu.

Change-Id: I738fe279933b4110cb31d0a3616aa4c52341c48a