Add page navigation tabs in correct order skin-side and remove js requirement for Vector tab icons
Tabs output in order they appear in the array in most skins, so just add the ones
we want at the beginning to the beginning of the array instead of the end.

Can't use oo-ui modules directly for the vector icons without js because they
require their classes be placed directly on the element, which is not possible
skin-side due to BaseTemplate::makeLink/makeListItem limitations, so we're just
copying the icons themselves out of oo-ui to apply ourselves. Also can't use
ResourceLoaderImageModule as PP does elsewhere because it doesn't work with
skinstyles, so instead of making a new module just to render the correct
versions of the icons or lumping it onto another with a bunch of other extra
stuff, we'll just manually make the pngs, apply the syles, etc...

Doesn't resolve lack of any navigation in Minerva, but should put it in basically
the correct order in all other skins, while also letting them style their own
icon tabs or whatever without an unnecessary js-dependency unless they actually
need one themselves for whatever reason.

Bug: T231250
Change-Id: Ifd8be387cef671b2aaa2a482c145bccf2e6d1543