Unset dynamic property on Parser when not needed any more
Dynamically creating a new class property on the Parser object is bad
enough. (Not uncommon, however.) But we should not leave it behind if
we don't need it any more.

The `?? false` is just a clever way of doing the same as before. If
the property is not set, it falls back to false.

Change-Id: I9d7fab2ea11d7815fd9d9ed59f4ca7d5ae06a917