Move trimming into PageContentBuilder::buildContentFromInput
This patch moves the rtrim() calls closer to the code where the
trimming actually matters. I feel the code responsible for trimming
user input should be the method that's named to take care of user
input: buildContentFromInput().

This also narrows the interface of the method down. Only the level
is needed, not the entire content object.

This patch also moves a few Phan suppressions closer to the lines of
code where the suppression is needed.

Change-Id: I0b46ea8c292f21612928c0e9d4bae57f3ba24555