Remove broken radio buttons from conflict resolution interface
When an edit on a page in the "Page:" namespace runs into a conflict,
core's EditPage class disables the custom UI where header, body, footer,
as well as the quality rank are separate UI elements. However, because
of the way the code is structured only half of the custom UI is disable.
This patch fixes this by disabling the rest as well.

I tried to fix the broken radio buttons by enabling the missing .js code.
This is pointless. The selection is not used. Instead, the quality rank
is already baked into the page content. This is used.

I tried to fix this the other way around and enable the custom UI
instead. This creates a very messy experience where the custom UI is
presented next to a raw wikitext field. After all, EditPage disables
custom UIs for a reason. It might still be possible to do it, but this
needs an actual product design.

Bug: T252863
Change-Id: I3a12ee90bd32ae6a9db202a13129db596eea565d