Fix loading animation not showing when dialog opens
The height of the container holding the image (in ImagePanel)
was not being set to any number while the image was hidden
causing flexbox to assume that the container had zero height,
hiding the animation. Thus, we set the height of the container
to 100% on load to show the animation.

Bug: T253458
Change-Id: I3fbe805270ebe046633f0a8c0bf2e9bdc32cd632