Use Openseadragon library for Pagelist Widget image zoom
Implemented zooming and panning for the Pagelist Widget
for PagelistWidget using the OpenSeadragon library.
Also added scripts to build the library for ProofreadPage
and check for changes.

This change should allow users to zoom and pan images
while using the pagelist widget which should improve the
user experience of users who work on scripts which
have a small font size.

Bug: T283917
Change-Id: I4e1d024e8c03d6666b9c9b2e1c0a57acb35acdde