Add prefetch links for prev and next page images
There are already prefetch links for the pages themselves,
so this adds similar links for the pages' images.

This also means scripts on the current page can access the URLs of
the prev and next images.

The $options array only ever has 'max-width', so I've moved that
inwards to the getImage() method. That key was also
being passed to $thumbnail->toHtml() but wasn't used there.

Lastly, the full-sized image URL is constructed in pretty much
the same way as the thumbnail URLs, so I've consolidated that in
the new method.

Bug: T230689
Change-Id: I192cb48b843f0e60d5990f2351f11c44905255c2