Use the WikiEditor ready hook instead of using() the lib
Just loading the library has the (unexpected?) side-effect of also
installing the toolbar. So respond to the existing
wikiEditor.toolbarReady hook to determine when to adjust the toolbar.

Don't enter the setup functions at all if a suitable toolbar is not
used, but do still initialise OSD so non-wikiEditor toolbar users can
still zoom.

Bug: T296033
Change-Id: I7fa03dd161f21f302253ef82e2f16261dde859ee