OSD: Prevent original image from getting overwritten by OSD
Previously, when switching layouts, destroying the current
OSD viewer instance would completely remove the original
image file. This could cause issues with scripts since
$( '.prp-page-image img' ) would return no elements if the
layout was switched.

The container for the OSD viewer has been moved to a child
of prp-page-image. This way, OSD will not remove the image
and old scripts should be able to access the original image
via $( '.prp-page-img img' )

Bug: T296153
Change-Id: Ie071632746a1ff8b52341e2b122ec9307dca8200