Allow to align page number in pagelist at right
Previously, hidden zeros allowed a coherent right
alignment for page numbers.

Per T277267, this solution has been replaced by the
proper use of span to indicate and allow to format
page numbers, avoiding strange spacing in other
contexts like in the JS code to offer a visual view
for pagelist.

This CSS change allows to style the page numbers,
to offer a presentation and alignment similar as
before.

Bug: T277267
Change-Id: If161cc1d99846c2ea3e3d341fe7c683cfcc8895d