Prevent PagelistWidget from rendering links inside buttons
The PagelistWidget used to consider all HTML data inside
the children of .index-pagelist to be part of the page number.
However, 19b3daf1d41ec17f13b196716199c41fc7b575df introduced a
new <span> wrapper to each and every page number, causing the
PagelistWidget to assume that the encapsulated link was also
part of the page number. Additionally, with the removal of the
leading zeros hack in the same commit, it does not make sense
for us to use the raw HTML as a page number.

This commit fixes the PagelistWidget's parsing of the generated
HTML and makes sure, that the PagelistWidget cannot include raw
HTML as part of the page number.

Bug: T277267
Change-Id: Id1bddd1cf1b2743fe3aa0364e6bd00bce33690c5