Remove limit on proofreadpagesinindex query API
The limit does not provide any performance benefit
since the API will internally have to load and cache
the pagelist data for all the pages anyway.

Additionally, adding the limit slows down users of the
API, cause large amounts of time being wasted
making requests back and forth for each of which anyway
loads all the pages via the pagelist.

Bug: T314022
Change-Id: Ide60c1d17700f44ecf5791cc26fe5a5c57789db3