modules: Rename internal foreign/ dir to lib/
Match what this directory tends to be called in other extensions
and in MediaWiki core.

Also move the ESLint ignores out to .eslintignore for faster
linting performance and for consistency with WikiLamba and other
extensions that have adopted T249858 meanwhile.

I've not migrated the exclude for `/modules/jquery/` as no such
directory exists. It was removed last year in commit 645dfcc92e
(Id6a7f66c0, T283917).

Bug: T249858
Change-Id: I2e6f1329224b98496ef7892fe19f865085e37d3c