Convert SetupAfterCache hook to HookHandler system
Changed the SetupAfterCache configuration in extension.json
and renamed the function initNamespaces to onSetupAfterCache in
ProofreadPageInit class to use the new HookHandlers system.
Refactored ProofreadPageInitTest and ProofreadPageTestCase.
Refactored hook function to load TemplateStylesNamespaces
if and only if the TemplateStyles extension is loaded.
Updated initNamespace to run before loading ProofreadPageNamespaceIds
and fixed linting errors in CI

Bug: T314911
Change-Id: Idd3c3e21146cc25b73c8efbd737e93cd01802d20