Set Proofread module targets explicitly
Currently ProofRead modules rely on whatever defaults are specified
in core (currently desktop site only). This default will change in
future so it's important we document which modules are meant to run
here.

In addition to this, we want to eventually drop the targets system
altogether, and allow all these modules to load on mobile.

This patch enables all modules where possible so we can address any
issues that surface directly.

Notable changes:
* Makes .pr_quality bar full width on mobile
* ext.proofreadpage.page.navigation is an empty module on Minerva
so harmless to load.
* Loads ext.proofreadpage.page styles

Note: ext.proofreadpage.icons is already targeting mobile as
that is the default for these modules so no update there is
required.

Bug: T324723
Change-Id: I26792696c41925740dbda6cf3b3a9b1a16f3c9b5