Prevent API (and EIS) from erroring if image is not generated
The imageforpages API errors out if it encounters a case where
the file exists but the thumbnail has not been rendered. In this
case, we should not send a URL and instead cancel prefetching for
the specific file. The image will get loaded normally when a user
navigates to the specific page.

NOTE: Calling the imagesforpage starts the rendering process so we
expect the image to be rendered by the time the user navigates to
the next page for which a thumbnail was not rendered.

Change-Id: If8181a7a8756a1e5d02cb1c50bfe52898cd7ca95