Bring page parsing implementation in line with backend
The previous DOMParser based parsing mechanism did not
account for badly formatted HTML markup causing a bunch
of superflous changes to wikitext markup. This has been
changed to be more inline with PageContentHandler::unserializeContentInWikitext(...)

NOTE: This approach does not support nested noincludes

This is a stop gap solution until T321446 is fixed with
eithier MCR or via some other method.

Change-Id: Ibaa877466690c03a638a865e79f84cc8f9b7787c