Avoid DB access in non-database tests
Prevent the database access with mocks or overrides when possible, and
add tests to the Database group otherwise.

PageContentHandlerTest::testGetParserOutputRedirectHtml was failing
after this change, but the test is way too coupled with the redirect
code in core. For instance, it already failed for me locally because the
URL scheme is different. It would also break with any slight change to
the relevant core logic. Instead, only check a small subset of the HTML;
this should be more than enough for this test, and it's less prone to
breaking.

Change-Id: I43e3b6b21303e782a492581a6a08e772503447be