PageContentHandlerTest: remove assignment to User::$mHideName
The two round-trip tests compare Content objects using assertEquals(),
which means that private properties of User objects are compared. This
fails when one of the User objects has cached values. The original
solution for this was to set a private property in User so that
hopefully nothing would be loaded. But that private property is going
away.

So, compare the objects using PageContent::equals().

Bug: T345683
Change-Id: I4898a23fcde34db8ef94b92d41722cedf9380dbc