Separate Vector and Vector-2022 styles for navigation
Fix the display of the prev/next/index navigtaion tabs in Vector-2022
by splitting the styles for the two skins into separate files.
This also changes the hover bottom border color to match the icon
rather than using #3366cc as is the current situation; this seems to be
in line with how the colors of the other links work.

This patch is basically Jdlrobson's work from Idb279f9b5692b7019a12ba8284795360c3811c76
and not really mine!
Thanks also to Punith.nyk for working on this issue.

Bug: T344665
Change-Id: I3d654bf0dfeb5e2da41ae23a7c3c6bc1a8a85beb