ext.proofreadpage.page.edit: fixes for CodeMirror 6 compatibility
In CM5, the CodeMirror library is loaded at runtime and not on page
load. This delay coincidentally gave ProofreadPage just enough time to
manipualte the DOM before CodeMirror is initialized. In reality, there
is a race condition that should be mitigated. With I4deeda192b we add a
hook in CM6, and make use of it here. CM5 still works due to the
aforementioned delay with mw.loader.using().

In CM5, the background color is set by codemirror.css in the CodeMirror
extension. CM6 uses the background color of the parent, which is why we
need to set it here.

These changes shouldn't effect CM5 integration.

Bug: T357794
Depends-On: I4deeda192bdc233101ba61739a636f8fd143c1de
Change-Id: I5c0824bf38cf7ad47a27be3c27d179cf2d940d7c