Hide link text for nav links, because images are used
Hide the inner text of the three navigation links for Page pages,
because the same text is already used for the `title` attribute.
Also add three new label messages for these links, to use as
`aria-label` values.

While we're here, fix the bottom border thickness and color to
match other nav links.

Bug: T362083
Change-Id: I0a2d98c5d01e61d66eee0e3c7451561fe2fb19a9