Allow non-english wikisources to set english numbering
For certain wikisources like Bengali Wikisource, the numbering in
pagelists defaults to the bengali numerals. However, if a book has
the English numeral system then these wikis have no way of specifing
that the numbering system is in English.

This patch adds a "latn" numbering scheme which corresponds to the
English numbering system which can be used by these wikis.

Bug: T371187
Change-Id: I6a0d7094c937f8a2bc1e05ce5afba0677a796430