Remove unnecessary call to ParserOutput::setPageProperty()
The default value for double-underscore magic words is `''`, and the
ParserAfterParse hook was being used to change it from `''` to `true`
but in fact the value was never used in practice (the query in
SpecialPagesWithoutScans doesn't use `pp_value` at all).  Remove a
deprecated use of boolean-valued page properties as well as an
unnecessary parser hook.

Change-Id: I2734383207b92f71bffc66ba2392a592a1df0954