page.edit, page.editinsequence: fix compatibility with CodeMirror 6
The older CodeMirror handler used the deprecated MutationEvent API to
detect changes in CodeMirror, which didn't work in CodeMirror 6. Check
if CM6 is enabled and if so, use the ext.CodeMirror.input hook since we
then don't need to be concerned on whether CM is enabled or not. The
onBodyChange() method uses jQuery.textSelection and thus doesn't need to
be editor-aware.

The CodeMirror 5 integration remains broken as of this patch due to an
unrelated bug in CodeMirror::Hooks. That can be fixed in a separate
patch, but of note is that the MutationEvent API has already been
removed from Chromium and Firefox is slated to follow, so the CM5
integration as written is doomed to break, anyway.

Also fix a race condition in ext.proofreadpage.page.edit.js. Sometimes
toggling CodeMirror on and off may set the textarea to the wrong size,
and this change should fix that.

Bug: T357794
Change-Id: I5132ec60cab292785be55f14f053cbb3dd747147