Lua: Prevent PHP errors in production from displayNumber lookup
Previously, if the following code was run onwiki:

```lua
local page = mw.ext.proofreadPage.newPage( 'Foobar.djvu/1' )
page.displayNumber
```

the corresponding PHP code would assume that if a Index page was found,
the associated page number for the Page: page would exist.

However, this is not always the case, in certain cases (such as if the
Index page is a redirect OR the Page: namespace page is out of bounds)
it is possible for a Index page to exist and not have the associated
number for the Page: namespace page being queried through Lua.

This patch adds a try {} catch {} around the code and returns a null
if the associated page is not found. This will prevent PHP production
errors and will instead push the error into a "normal" Lua error on
misbehaving template allowing for wikis to deal with it instead.

Bug: T383924
Change-Id: I32b31f5c797a81b275f628c5af2789455a36b340
(cherry picked from commit 19e7e0b65ddb6f0622ee3c7125c83f454ea45737)