ext.proofreadpage.page.edit: add CodeMirror to header and footer
Create a new CodeMirror subclass for use on the header and footer. Add a
setupCodeMirror() method to the edit module, which then lazy loads the
new class and creates instances for the header and footer.

For now search is confined to the main textarea, which may not be a huge
concern given the header/footer are generally quite small to begin with.

CodeMirror preferences (accessible via Mod-Shift-, or the settings icon
in the Advanced section of the WikiEditor toolbar) are controlled only
by the main CodeMirror instance and synced to the header/footer. The
preferences UI is awkwardly confined to the main textarea. This will
likely eventually become a document-wide modal in the future, so in due
time it will be not-so-awkward.

When focus enters one of the textareas, the WikiEditor context is
updated to point to that textarea so that the toolbar buttons act where
expected. This is labeled as a HACK as it might come with unintended
side effects.

Bug: T380262
Change-Id: Ie937f4e8cb356e402807837493e5981f6120ae42