Only load VisualEditor integration when it is loaded
On CI, we would like to remove the recursive injection of extensions
dependencies. When ProofreadPage is a dependency, we are required to add
VisualEditor to fullfil the requirement that all dependencies are
defined (ResourcesTest::testValidDependencies).

Move the `ext.proofreadpage.ve.node.pages` registration from
`extension.json` to the PHP `ResourceLoaderRegisterModules`. Only load
it when VisualEditor is loaded.

Bug: T389998
Change-Id: I0dbed826b7e486eae1143fed87737ecec9ec2123