Wrap pages output with div after wikitextToDOM call
* Avoid issues with top level mw:Transclusion
* Add test to reproduce the rt-testing failure

When the page referenced in the <pages> has an unopened div,
the wikitextToDom ends up tagging the firstChild of the
document fragment as mw:Transclusion because of the unbalanced
tree causing the error.

By wrapping the returned document fragment after wikitextToDom
the top level element is not tagged as mw:Transclusion allowing
parsoid to render the tag properly.

Bug: T408222
Change-Id: I73976889387eb169a29fb9ad1ce045f3a8d8e3bc